Update methods in cctpr-evm
