+++ packages/cctp-sdk/cctpr-evm/src/registry.ts
@@ -2,9 +2,14 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
+import type { TransferGeneratorT, TransferGeneratorTReturn, TransferOptions } from "./transfer.js";
 
 declare module "@stable-io/cctp-sdk-cctpr-definitions" {
   export interface PlatformImplsOf {
-    Evm: never;
+    Evm: {
+      TransferOptions: TransferOptions;
+      TransferGeneratorT: TransferGeneratorT;
+      TransferGeneratorTReturn: TransferGeneratorTReturn;
+    };
   }
 }
